tested usability of global light uniforms
More tests are needed for camera rotation, etc..

+++ examples/vert_color_norm.glsl
@@ -1,9 +1,9 @@
 in highp vec4 position;
 in mediump vec4 color;
-in mediump vec4 normal;
+in highp vec4 normal;
 
 out highp vec4 color_v;
-out mediump vec3 normal_v;
+out highp vec3 normal_v;
 out mediump vec3 eye_v;
 
 void main()
